Replaced azcopy (which caused an authentication error) with Hugging Face Hub for NOTSOFAR data download.
