+++ requirements.txt
@@ -6,4 +6,4 @@ click==8.1.7
 torchaudio>=0.13.1
 regex==2023.12.25
 more-itertools==10.2.0
-azure-cli==2.57.0
+huggingface-hub==0.35.0
